+++ components/faphub/dao/network/src/main/java/com/flipperdevices/faphub/dao/network/ktorfit/model/detailed/KtorfitSourceCodeLinks.kt
@@ -1,0 +1,10 @@
+package com.flipperdevices.faphub.dao.network.ktorfit.model.detailed
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class KtorfitSourceCodeLinks(
+    @SerialName("type") val type: String,
+    @SerialName("uri") val url: String
+)