+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - [Feature] Add report app button
 - [Feature] Now the installed apps are even shown offline
 - [Feature] Change design for installed tab
+- [Feature] Add more error catching for FapHub
 - [Refactor] Migrate all feature components modules from KSP to Anvil
 - [Refactor] Migrate to Ktorfit
 - [Refactor] Migrate key emulate to new module, decompose Emulate Helper